Eliminate circular dependency
Moved Decoder abstract class into its own DecoderClass.ts file
