+++ metar-lookup/src/decoder/CloudDecoder.ts
@@ -1,4 +1,4 @@
-import { Decoder } from "./DecoderUtils";
+import { Decoder } from "./DecoderClass";
 
 export class CloudDecoder extends Decoder{
