+++ metar-lookup/src/decoder/DecoderClass.ts
@@ -1,0 +1,9 @@
+export abstract class Decoder {
+    decodedText: string
+
+    constructor(decodedText: string) {
+        this.decodedText = decodedText;
+    }
+
+    abstract Decode(raw: string): Array<string|boolean>;
+}
