+++ metar-lookup/src/decoder/TemperatureDecoder.ts
@@ -1,4 +1,4 @@
-import { Decoder } from "./DecoderUtils";
+import { Decoder } from "./DecoderClass";
 
 export class TemperatureDecoder extends Decoder {
